moved rank_corr compare to tests exclusively, now superseeded by compare_spearman
